+++ docs/client/components/pages/Counter/CustomCounterEditor/index.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import { List } from 'immutable';
 import Editor, { createEditorStateWithText } from 'draft-js-plugins-editor'; // eslint-disable-line import/no-unresolved
 import createCounterPlugin from 'draft-js-counter-plugin'; // eslint-disable-line import/no-unresolved
 import editorStyles from './editorStyles.css';
@@ -10,7 +11,7 @@ const theme = {
 };
 const counterPlugin = createCounterPlugin({ theme });
 const { CharCounter, WordCounter, LineCounter, CustomCounter } = counterPlugin;
-const plugins = [counterPlugin];
+const plugins = List([counterPlugin]);
 const text = `This editor has counters below!
 Try typing here and watch the numbers go up. 🙌