+++ docs/client/components/pages/Sticker/CustomStickerEditor/index.js
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import { List } from 'immutable';
 import { EditorState } from 'draft-js';
 import Editor from 'draft-js-plugins-editor'; // eslint-disable-line import/no-unresolved
 import createStickerPlugin from 'draft-js-sticker-plugin'; // eslint-disable-line import/no-unresolved
@@ -6,7 +7,7 @@ import editorStyles from './editorStyles.css';
 import stickers from './stickers';
 
 const stickerPlugin = createStickerPlugin({ stickers });
-const plugins = [stickerPlugin];
+const plugins = List([stickerPlugin]);
 const StickerSelect = stickerPlugin.StickerSelect;
 
 export default class CustomStickerEditor extends Component {